+++ src/main/java/ar/edu/itba/montu/abstraction/LocatableAgent.java
@@ -25,7 +25,7 @@ public abstract class LocatableAgent extends Agent{
 		return kingdom;
 	}
 
-	public abstract void defend(LocatableAgent agent, int harm);
+	public abstract void defend(LocatableAgent agent, double harm);
 
 	public abstract boolean isAlive();
